requirements.txt
streamlit
groq
python-dotenv
SpeechRecognition
pydub
pyaudio
